+++ CubeGame/World/Constants.h
@@ -4,6 +4,6 @@
 
 const unsigned int CHUNK_SIZE = 16;
 
-const uint8_t AIR_BLOCK = 0;
-const uint8_t DIRT_BLOCK = 1;
-const uint8_t GRASS_BLOCK = 2;
+const int AIR_BLOCK = 0;
+const int DIRT_BLOCK = 1;
+const int GRASS_BLOCK = 2;
